Remove `Service::class` model
