+++ database/factories/ProviderFactory.php
@@ -5,7 +5,7 @@ namespace Payavel\Orchestration\Database\Factories;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 use Payavel\Orchestration\Models\Provider;
-use Payavel\Orchestration\Models\Service;
+use Payavel\Orchestration\Service;
 
 class ProviderFactory extends Factory
 {
@@ -39,11 +39,9 @@ class ProviderFactory extends Factory
     {
         return $this->afterMaking(function (Provider $provider) {
             if(is_null($provider->service_id)) {
-                $service = Service::inRandomOrder()->firstOr(
-                    fn () => Service::factory()->create()
-                );
-
-                $provider->service_id = $service->id;
+                $provider->service_id = is_null($service = Service::all()->random())
+                    ? Str::lower($this->faker->unique()->word())
+                    : $service->getId();
             }
 
             $studlyProvider = Str::studly($provider->id);
