+++ tests/Feature/Console/Commands/TestOrchestrateProviderCommand.php
@@ -7,10 +7,12 @@ use Payavel\Orchestration\Service;
 use Payavel\Orchestration\Tests\Contracts\CreatesServiceables;
 use Payavel\Orchestration\Tests\TestCase;
 use Payavel\Orchestration\Tests\Traits\AssertsGatewayExists;
+use Payavel\Orchestration\Tests\Traits\CreatesServices;
 
 abstract class TestOrchestrateProviderCommand extends TestCase implements CreatesServiceables
 {
-    use AssertsGatewayExists;
+    use AssertsGatewayExists,
+        CreatesServices;
 
     /** @test */
     public function make_payment_provider_command_will_prompt_for_missing_arguments()
