+++ tests/Traits/CreatesConfigServiceables.php
@@ -15,24 +15,6 @@ use Payavel\Orchestration\DataTransferObjects\Service;
 trait CreatesConfigServiceables
 {
     /**
-     * Creates a serviceable instance.
-     *
-     * @param array $data
-     * @return \Payavel\Orchestration\Contracts\Serviceable
-     */
-    public function createService($data = [])
-    {
-        $data['id'] = $data['id'] ?? Str::lower($this->faker->unique()->word());
-        $data['test_gateway'] = $data['test_gateway'] ?? '\\App\\Services\\' . Str::studly($data['id']) . '\\Fake' . Str::studly($data['id']) . 'Request';
-
-        Config::set('orchestration.services.' . $data['id'], $serviceSlug = Str::slug($data['id']));
-
-        Config::set($serviceSlug . '.testing.gateway', $data['test_gateway']);
-
-        return new Service($data);
-    }
-
-    /**
      * Creates a providable instance.
      *
      * @param Serviceable|null $service
@@ -91,35 +73,6 @@ trait CreatesConfigServiceables
                 Config::get($providers, []),
                 [$provider->getId() => $data]
             )
-        );
-    }
-
-    /**
-     * Sets the default configuration for a serviceable instance.
-     *
-     * @param Serviceable $service
-     * @param Merchantable|null $merchant
-     * @param Providable|null $provider
-     * @return void
-     */
-    public function setDefaultsForService(Serviceable $service, Merchantable $merchant = null, Providable $provider = null)
-    {
-        Config::set(
-            Str::slug($service->getId()) . '.defaults.merchant',
-            $merchant instanceof Merchantable ? $merchant->getId() : $merchant
-        );
-
-        if (is_null($provider) && ! is_null($merchant)) {
-            $provider = Collection::make(
-                Config::get(Str::slug($service->getId()) . '.merchants.' . $merchant->getId() . '.providers')
-            )
-                ->keys()
-                ->first();
-        }
-
-        Config::set(
-            Str::slug($service->getId()) . '.defaults.provider',
-            $provider instanceof Providable ? $provider->getId() : $provider
         );
     }
 }
