+++ tests/Traits/CreatesDatabaseServiceables.php
@@ -7,21 +7,9 @@ use Payavel\Orchestration\Contracts\Providable;
 use Payavel\Orchestration\Contracts\Serviceable;
 use Payavel\Orchestration\Models\Merchant;
 use Payavel\Orchestration\Models\Provider;
-use Payavel\Orchestration\Models\Service;
 
 trait CreatesDatabaseServiceables
 {
-    /**
-     * Creates a serviceable instance.
-     *
-     * @param array $data
-     * @return \Payavel\Orchestration\Contracts\Serviceable
-     */
-    public function createService($data = [])
-    {
-        return Service::factory()->create($data);
-    }
-
     /**
      * Creates a providable instance.
      *
@@ -71,27 +59,5 @@ trait CreatesDatabaseServiceables
         throw_unless($merchant instanceof Merchant);
 
         $merchant->providers()->sync([$provider->getId() => $data], false);
-    }
-
-    /**
-     * Sets the default configuration for a serviceable instance.
-     *
-     * @param Serviceable $service
-     * @param Merchantable|null $merchant
-     * @param Providable|null $provider
-     * @return void
-     */
-    public function setDefaultsForService(Serviceable $service, Merchantable $merchant = null, Providable $provider = null)
-    {
-        throw_unless($service instanceof Service);
-
-        if (is_null($provider) && ! is_null($merchant)) {
-            $provider = $merchant->default_provider_id;
-        }
-
-        $service->update([
-            'default_merchant_id' => $merchant instanceof Merchantable ? $merchant->getId() : $merchant,
-            'default_provider_id' => $provider instanceof  Providable ? $provider->getId() : $provider,
-        ]);
     }
 }
